feat: add extract_anatomy helper
